Remove DOCKER_BUILDKIT=1 dependency by fixing COPY
Using ../ for the left side of COPY was a mistake that was present in
this repo for a long time.

Buildkit might have silently adjusted the path to be . so we never go
back beyond the build context.

In any case, it's fixed now. If you're using Buildkit it makes no
difference but now the project works without it but more importantlt
the correct path is being used.